added createPartitions() api

+++ clients/src/main/java/org/oracle/okafka/clients/consumer/internals/AQKafkaConsumer.java
@@ -263,7 +263,7 @@ public final class AQKafkaConsumer extends AQClient{
 				else {
 					log.debug("Received other than AQjmsBytesMessage");
 					try {
-						int partition = messages[i].getIntProperty(AQClient.PARTITION_PROPERTY);
+						int partition = (int) messages[i].getLongProperty(AQClient.PARTITION_PROPERTY);
 						OKafkaOffset  okOffset = 	MessageIdConverter.getOKafkaOffset(messages[i].getJMSMessageID(), true, true);
 						long offset = okOffset.getOffset();
 						log.error("Message is not an instance of AQjmsBytesMessage: Topic {} partition {} offset{}",topic, partition, offset );
